Fixed issues in processing of large DL packets [Issue-ID: ODUHIGH-309]

+++ src/5gnrsch/sch_utils.h
@@ -88,7 +88,7 @@ int8_t searchSpaceIdxTable[MAX_SEARCH_SPACE_INDEX][4];
 
 /* functions declarations */
 void freqDomRscAllocType0(uint16_t startPrb, uint16_t prbSize, uint8_t *freqDomain);
-uint16_t schCalcTbSize(uint16_t payLoadSize);
+uint16_t schCalcTbSize(uint32_t payLoadSize);
 uint16_t schCalcNumPrb(uint16_t tbSize, uint16_t mcs, uint8_t numSymbols);
 uint16_t schCalcTbSizeFromNPrb(uint16_t numPrb, uint16_t mcs, uint8_t numSymbols);
 SchUeCb* schGetUeCb(SchCellCb *cellCb, uint16_t crnti);
